refactor(types): move constnat data and types
