+++ src/app/lib/types.ts
@@ -5,3 +5,7 @@ export type NewsDataFormatted = {
   story_url: string
   story_id: string
 }
+export type NewsType = {
+  value: string
+  label: JSX.Element
+}
