--- src/app/utils/home/constant.ts
@@ -1,7 +1,0 @@
-export const LIMIT = 8
-export const TAB_ALL = 'all'
-export const TAB_FAVES = 'faves'
-export const TABS = [
-  { title: 'All', value: TAB_ALL },
-  { title: 'My faves', value: TAB_FAVES },
-]
